# Spark 3.2: Optimized add_files procedure's listPartitions