AddFilesProcedure uses InMemoryFileIndex to get SparkPartitions. InMemoryFileIndex is not used efficiently as it lists _all_ the files in a given path to discover partitions which is memory intensive and not really needed as we just want to find the partitions. This can be achieved without listing all the files

This PR replaces InMemoryFileIndex  with a custom implementation of _PartitioningAwareFileIndex_ which avoids listing _all_ files and only lists the directories to discover partitions. 

This optimization reduces _listPartitions_ _latency_ by > 90% and _memory_ usage more than 3 times. 

Addresses Issue: https://github.com/apache/iceberg/issues/7027

Tested with input S3 folder with large number of files.
**All Files in path**: 7,169,419
**Total size**: 6.7 TB

Filter on date and hour, where single partition has 41,417 files

| Cluster Config  | Before |  After     |  Improvement     |
| ------------- | ------------- | ------------- | ------------- |
| Driver Memory | 128GB (Fails with OOM if < 128GB)  | 8GB | ~93% |
| Executor Memory  | 64GB  | 8GB | ~87% |
| Num of Executors  | 20 |4  | ~80% |
| List Partitions Latency  | 12M49.459S  | 0.774S  | ~99% |

Latency to add the files in all partitions once they are listed is the same, before 2H25M14.905S and after 2H26M15.803S.

### Testing
Existing tests cover various use cases and they all pass with new implementation.
Did load and performance tests with medium and large S3 paths.