@abmo-x 2023-04-18
Had to create a Scala class as this PR is extending a base Scala class in Spark. When I created it as Java class compilation fails with Scala 2.13