@ajantha-bhat 2023-12-14
Closing this as Spark 3.2 is no longer supported in the latest version.

If issue exist on other spark version, please handle there.